@alexey-milovidov 2022-05-08
The improvement is only marginal.

But there are the following caveats:
1. This code path will be run only if ClickHouse is compiled with AVX-512 globally enabled. But we don't do that, because AVX-512 is rare. And we don't support any builds of ClickHouse except the single official build.
2. Implementing dynamic dispatching for this code will be difficult if not impossible.
3. The changes have been tested on unknown CPU model, but the performance of AVX-512 can vary a lot on different CPU models and on some models it can lead to performance degradation.
4. We add 15 byte padding to the columns, but with the introduction of AVX-512, it will be reasonable to change the padding to 63 bytes and simplify the code even more. It was not done in this PR.
5. It has been tested on TPC-H or Star Schema Benchmark dataset that is known to represent unrealistic workload with too many pseudorandom data with unnatural distributions. It may lead to excessive copying and overly-optimistic results on this benchmark.
6. Some prior work https://github.com/ClickHouse/ClickHouse/pull/21812 and https://presentations.clickhouse.com/zero_cost_conf/memcpy/ advices against this optimization.