@yaqi-zhao 2022-05-10
Thanks for your comments. @alexey-milovidov  It is useful and helps us provide more comprehensive solution in the future. But there are some confusions about the above caveats;
1.	I found that prior work on avx512 optimization have been merged, e.g.,,Avx512 memcmp small [#33706](https://github.com/ClickHouse/ClickHouse/pull/33706) && add x86 feature avx2/avx512 support for accelerating filter operation [#30014](https://github.com/ClickHouse/ClickHouse/pull/30014). But the comment said ck don't support any builds of ClickHouse except the single official build. I have doubt that if we have more avx512 optimization in the future, can these features be merged?
2.	Comments say “The improvement is only marginal” --=>  Is that means 5% performance improvement on function query tested by Star Schema Benchmark to small? Generally, how much the improvement made will the PR be accepted?