@alexey-milovidov 2022-05-14
5% is good if it is available in the default build configuration and not good to motivate some code path that works only in custom builds.