@yaqi-zhao 2022-05-16
@alexey-milovidov  what if I add a cmake option to enable the code by default if build enviromnent supports AVX512?