@alexey-milovidov 2022-05-17
It's not a good option, we can build on servers with AVX-512 but deploy on other servers.
Dynamic dispatch will be better if they can really work on memcpySmall (that's in question).